requirements.txt
yfinance>=0.2.18
openai>=1.0.0
anthropic>=0.20.0
groq>=0.4.0
schedule>=1.2.0
alpaca-trade-api>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pytz>=2023.3
